- Renaming files in preparation for rails 3.0.0.beta3
